requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
requests==2.31.0
python-telegram-bot==13.7
python-dotenv==1.0.0